Fixed Out of Bounds error for expertise, changed modifier on skillmods to float
(BugFix) Fixed out of bounds error when assigning expertise points
(Code) Changed SkillMod's modifier variable to float value
(Code) Changed Charge! and Scatter! to use updated addBuffToCreature

+++ src/resources/objects/creature/CreatureMessageBuilder.java
@@ -163,7 +163,7 @@ public class CreatureMessageBuilder extends ObjectMessageBuilder {
 				buffer.put((byte) 0);
 				buffer.put(getAsciiString(skillMod.getName()));
 				buffer.putInt(skillMod.getBase());
-				buffer.putInt(skillMod.getModifier());
+				buffer.putInt((int) skillMod.getModifier());
 			}
 		}
 		buffer.putFloat(creature.getSpeedMultiplierBase());
